Build ONNX Runtime from a submodule, and fix the warnings that were ours
Every dependency is now a pinned submodule. ONNX Runtime was the exception — a
SHA256-pinned prebuilt archive — and is now libs/onnxruntime at v1.23.2, built from
source through ExternalProject with only the CPU provider and only the shared library
installed.

Three arguments turned out to be mandatory, all because the pinned release is older
than the toolchain compiling it. --compile-no-warning-as-error, because ONNX Runtime
sets COMPILE_WARNING_AS_ERROR on every target and GCC 16 rejects 1.23.2 outright over
-Wsfinae-incomplete and -Wdangling-pointer. CMAKE_POLICY_VERSION_MINIMUM=3.5, because
CMake 4 refuses the pre-3.5 minimums its fetched dependencies declare. And
-include cstdint on GCC 15+, for core/common/semver.h, which uses uint32_t without
including it — the one place the sources genuinely are wrong, fixed with a flag rather
than by patching a pinned submodule.

Two costs are real and are documented rather than hidden: Python 3 is required as ONNX
Runtime's own build tool, since its CMake generates the export symbol list by running
gen_def.py, and its configure still fetches abseil, protobuf, flatbuffers and the rest
from its deps.txt. CI reads the pinned commit with `git rev-parse HEAD:libs/onnxruntime`
and caches the install tree against it, so a hit skips both the 900 MB clone and the
build by pointing RVCARA_ONNXRUNTIME_ROOT at what was restored.

Verified from scratch: configure, ONNX Runtime built and installed, plug-in built,
28/28 tests pass, the VST3 dlopens and carries its ARA main factory class. Converting a
sung phrase through the from-source library agrees with the prebuilt at correlation
0.999937 — not bit-identical, because MLAS selects kernels for the machine it compiled
on — and runs marginally faster, 0.62x real time against 0.65x.

The warning tiers are re-cut by ownership instead of by the system-include promotion of
the previous commit, which silenced JUCE rather than judging us. Measured first: of 6,284
warnings in a full build, 6,282 were inside JUCE's own headers and none were in our code.
The strict tier — now also -Wconversion, -Wsign-conversion, -Wuseless-cast, -Wcast-qual,
-Wextra-semi — is applied to the translation units that include nothing from libs/, and
the DSP core is kept framework-free so that it can carry it. Every source here was then
checked by hand under the full set with the output filtered to src/ and tests/, which
found the three defects JUCE's noise had been burying:

- PluginProcessor::processBlock(float) hid the double-precision overload. Nothing calls
  it today, but a hidden virtual is how a host silently gets a no-op.
- A useless cast on MemoryMappedFile::getSize(), which already returns size_t.
- Comments pointing at tools/rvcara_export, deleted two commits ago.

Nothing is suppressed, so the honest total for a full build is 1,054 warnings: 966 in
ONNX Runtime's sources, 85 in the ONNX schema library, 1 in abseil, 2 in JUCE's vendored
harfbuzz, and none in RVCARA.

+++ src/BinaryMatrix.cpp
@@ -39,7 +39,7 @@ BinaryMatrix BinaryMatrix::load (const juce::File& file)
         return matrix;
     }
 
-    const auto numBytes = static_cast<std::size_t> (mapping->getSize());
+    const auto numBytes = mapping->getSize();
 
     if (numBytes < headerSize)
     {
